requirements.txt
# Python dependencies for Job Application Generator

# LLM API
anthropic>=0.40.0

# PDF processing
PyPDF2>=3.0.0

# Web scraping
requests>=2.31.0
beautifulsoup4>=4.12.0

# Markdown to PDF conversion
markdown>=3.5.0
weasyprint>=60.0

# Environment variables
python-dotenv>=1.0.0
